fix: внести изменения согласно предложениям в PR
