+++ src/components/Statistic/Statistic.module.scss
@@ -3,6 +3,7 @@
 div {
   &:global(.ant-statistic-title) {
     color: white;
+    text-align: left;
   }
   &:global(.ant-statistic-content) {
     color: white;
