+++ src/pages/Profile/components/FormControls/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FormControls.view";
