+++ src/pages/Profile/components/FormFields/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FormFields.view";
