+++ src/pages/Profile/components/UploadButton/UploadButton.module.scss
@@ -1,0 +1,8 @@
+.upload {
+  color: white;
+}
+
+.button {
+  margin-top: 8px !important;
+  font-size: 9px;
+}
