+++ src/pages/Profile/components/UploadButton/UploadButton.view.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import { PlusOutlined } from "@ant-design/icons";
+
+import styles from "./UploadButton.module.scss";
+
+const UploadButton = () => {
+  return (
+    <div className={styles.upload}>
+      <PlusOutlined />
+      <div className={styles.button}>Загрузить</div>
+    </div>
+  );
+};
+export default UploadButton;
