+++ src/pages/Profile/components/UploadButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UploadButton.view";
